requirements.txt
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
redis==5.0.8
google-generativeai==0.8.2
google-api-python-client==2.143.0
google-auth-oauthlib==1.2.1
httpx==0.27.2
Pillow==10.4.0
APScheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
aiofiles==24.1.0
python-multipart==0.0.12
pytz==2024.1
